backend/ai_service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
google-generativeai
pymongo
python-dotenv
requests
